+++ part3/controllers/user.js
@@ -1,4 +1,6 @@
 const userService = require('../services/user');
+const tokenService = require('../services/token');
+
 
 const createUser = async (req, res) => {
     const user = await userService.createUser(req.body.name, req.body.password, req.body.email, req.body.image);
@@ -10,11 +12,21 @@ const createUser = async (req, res) => {
 };
 
 const getUsers = async (req, res) => {
+    const headersUser = await tokenService.validateHeadersUser(req.headers['user']);
+    if (!headersUser)
+    {
+        return res.status(400).json({ errors: ['Header User doesn\'t exist'] });
+    }
     const users = await userService.getUsers();
     res.json(users);
 };
 
 const getUser = async (req, res) => {
+    const headersUser = await tokenService.validateHeadersUser(req.headers['user']);
+    if (!headersUser)
+    {
+        return res.status(400).json({ errors: ['Header User doesn\'t exist'] });
+    }
     const user = await userService.getUserById(req.params.id);
     if(!user){
         return res.status(404).json({ errors: ['User not found'] });
@@ -23,6 +35,11 @@ const getUser = async (req, res) => {
 };
 
 const updateUser = async (req,res) => {
+    const headersUser = await tokenService.validateHeadersUser(req.headers['user']);
+    if (!headersUser)
+    {
+        return res.status(400).json({ errors: ['Header User doesn\'t exist'] });
+    }
     const updatedUser = await userService.updateUser(req.params.id, req.body.name, req.body.password, req.body.email, req.body.image, req.body.movie);
     if(!updatedUser){
         return res.status(404).json({ errors: ['User not found'] });
@@ -31,6 +48,11 @@ const updateUser = async (req,res) => {
 };
 
 const deleteUser = async (req,res) => {
+    const headersUser = await tokenService.validateHeadersUser(req.headers['user']);
+    if (!headersUser)
+    {
+        return res.status(400).json({ errors: ['Header User doesn\'t exist'] });
+    }
     const deletedUser = await userService.deleteUser(req.params.id);
     if(!deleteUser){
         return res.status(404).json({ errors: ['User not found'] });
